Fix workers to use handle_call
Don't cast pooled workers, really. A worker is checked back in to the
pool immediately after casting on it, and will be checked out first
among other workers due to the LIFO strategy. The worker that checked
out will certainly be running its task, so succeeding casts will pile up
on the inbox of that specific worker.